Fixed port downcast issue

+++ sln/Amongst/src/MongoDBConnection.cs
@@ -5,11 +5,11 @@ namespace Amongst
     public class MongoDBConnection
     {
         public IPAddress IP { get; }
-        public short Port { get; }
+        public int Port { get; }
 
         //------------------------------------------------------------------------------------------------------------->
 
-        public MongoDBConnection(IPAddress ip, short port)
+        public MongoDBConnection(IPAddress ip, int port)
         {
             IP = ip;
             Port = port;
